fix: Add missing extension and size fields

+++ app/src/main/java/gateway/services/ServiceMetadata.java
@@ -102,7 +102,7 @@ public class ServiceMetadata
 		public String name;
 		public String extension;
 		public Integer volume;
-		public long size;
+		public int size;
 		public boolean isShared;
 	}
 
@@ -130,7 +130,7 @@ public class ServiceMetadata
 				s.name = resBody.getString ("name");
 				s.extension = resBody.isNull ("extension") ? null : resBody.getString ("extension");
 				s.volume = resBody.isNull ("volume") ? null : resBody.getInt ("volume");
-				s.size = resBody.getLong ("size");
+				s.size = resBody.getInt ("size");
 				s.isShared =
 					resBody.getBoolean ("is_shared"); // TODO make a test for true and false
 			} else {
